Added test quick sort

+++ matrix.h
@@ -27,6 +27,7 @@ typedef struct {
  * @return Matrix* 
  */
 Matrix *new_matrix(unsigned int size_n, unsigned int size_m);
+
 /**
  * @brief Deletes the given Matrix from memory.
  * 
@@ -47,6 +48,7 @@ void delete_matrix(Matrix *matrix);
 void resize(unsigned int new_size_n, unsigned int new_size_m,
             unsigned int *row_indexes, unsigned int *col_indexes,
             Matrix *matrix);
+            
 /**
  * @brief Multiplies two given Matrices and returns the new one.
  * 
@@ -55,6 +57,7 @@ void resize(unsigned int new_size_n, unsigned int new_size_m,
  * @return Matrix* 
  */
 Matrix *multiply_matrices(Matrix *matrix_a, Matrix *matrix_b);
+
 /**
  * @brief Multiplies two given Matrices using Strassen's algorithm and returns a
  *        new one.
@@ -64,12 +67,14 @@ Matrix *multiply_matrices(Matrix *matrix_a, Matrix *matrix_b);
  * @return Matrix* 
  */
 Matrix *strassen_multiply_pot_matrices(Matrix *matrix_a, Matrix *matrix_b);
+
 /**
  * @brief Transposes the given Matrix and saves it at the same address.
  * 
  * @param matrix 
  */
 void transpose_matrix(Matrix *matrix);
+
 /**
  * @brief Interface for raising the given Matrix to a power in logarithmic time
  *        and saves it at the same address.
@@ -79,6 +84,7 @@ void transpose_matrix(Matrix *matrix);
  */
 void power_raise_matrix(unsigned int power, Matrix **matrix);
 
+
 /**
  * @brief Adds two Matrices and returns the result as a new Matrix.
  * 
@@ -87,6 +93,7 @@ void power_raise_matrix(unsigned int power, Matrix **matrix);
  * @return Matrix* 
  */
 Matrix *add_matrices(Matrix *a, Matrix *b);
+
 /**
  * @brief Subtracts two Matrices and returns the result as a new Matrix.
  * 
@@ -95,6 +102,7 @@ Matrix *add_matrices(Matrix *a, Matrix *b);
  * @return Matrix* 
  */
 Matrix *subtract_matrices(Matrix *a, Matrix *b);
+
 /**
  * @brief Swaps two Matrices.
  * 
@@ -102,6 +110,7 @@ Matrix *subtract_matrices(Matrix *a, Matrix *b);
  * @param b 
  */
 void swap_matrices(Matrix **a, Matrix **b);
+
 /**
  * @brief Resizes the given Matrix, ignoring the old values.
  * 
@@ -111,6 +120,7 @@ void swap_matrices(Matrix **a, Matrix **b);
  */
 void resize_matrix(unsigned int new_size_n, unsigned int new_size_m,
                    Matrix *matrix);
+
 /**
  * @brief Breaks the given Matrix in 4 equal square blocks. Returns through parameters.
  * @note The specified matrix must have equal sizes, powers of two.
@@ -123,6 +133,7 @@ void resize_matrix(unsigned int new_size_n, unsigned int new_size_m,
  */
 void break_matrix_in_blocks(Matrix *matrix, Matrix **a, Matrix **b, Matrix **c,
                             Matrix **d);
+
 /**
  * @brief Builds a Matrix from 4 smaller blocks. Returns a new Matrix.
  * @note The building blocks must all have equal sizes, powers of two.
@@ -134,6 +145,7 @@ void break_matrix_in_blocks(Matrix *matrix, Matrix **a, Matrix **b, Matrix **c,
  * @return Matrix* 
  */
 Matrix *build_matrix_from_blocks(Matrix *a, Matrix *b, Matrix *c, Matrix *d);
+
 /**
  * @brief The implementation of logarithmic power raising. Returns result Matrix as parameter.
  * 
@@ -143,6 +155,7 @@ Matrix *build_matrix_from_blocks(Matrix *a, Matrix *b, Matrix *c, Matrix *d);
  */
 void recursive_power_raise_matrix(unsigned int power, Matrix *matrix,
                                   Matrix **result_matrix);
+
 /**
  * @brief Calculates the sum of all the elements of the matrix.
  * @note The sum is modulo.
@@ -151,26 +164,28 @@ void recursive_power_raise_matrix(unsigned int power, Matrix *matrix,
  * @return int
  */
 int compute_elements_sum(const Matrix *matrix);
+
 /**
- * @brief Compares sum of Matrices. If a < b returns 1, 
+ * @brief Calculates the difference between Matrix a and Matrix b;
  *        if a > b returns -1, else returns 0.
  * @note Used for sorting algorithms.
  * 
  * @param a 
  * @param b 
- * @return short int 
+ * @return int 
  */
-short int cmp_matrices_ascending(const Matrix *a, const Matrix *b);
+int cmp_matrices_ascending(const Matrix *a, const Matrix *b);
+
 /**
- * @brief Compares sum of Matrices. If a > b returns 1,
- *        if a < b returns -1, else returns 0.
+ * @brief Calculates the difference between Matrix b and Matrix a;
  * @note Used for sorting algorithms.
  * 
  * @param a 
  * @param b 
- * @return short int 
+ * @return int 
  */
-short int cmp_matrices_descending(const Matrix *a, const Matrix *b);
+int cmp_matrices_descending(const Matrix *a, const Matrix *b);
+
 /**
  * @brief Applies modulo on given value and returns new value.
  * @note Used for overflow protection.
